+++ src/Agent/NewRelic/Agent/Core/Segments/DatastoreSegmentData.cs
@@ -170,7 +170,7 @@ namespace NewRelic.Agent.Core.Segments
 
         public override string GetTransactionTraceName()
         {
-            var name = Model == null ? DatastoreVendorName.GetDatastoreOperation(Operation) : MetricNames.GetDatastoreStatement(DatastoreVendorName, Model, Operation);
+            var name = string.IsNullOrEmpty(Model) ? DatastoreVendorName.GetDatastoreOperation(Operation) : MetricNames.GetDatastoreStatement(DatastoreVendorName, Model, Operation);
             return name.ToString();
         }
 
@@ -213,7 +213,11 @@ namespace NewRelic.Agent.Core.Segments
                 AttribDefs.DbStatement.TrySetValue(attribVals, GetObfuscatedSQL());
             }
 
-            AttribDefs.DbCollection.TrySetValue(attribVals, _parsedSqlStatement.Model);
+            if (!string.IsNullOrEmpty(_parsedSqlStatement.Model))
+            {
+                AttribDefs.DbCollection.TrySetValue(attribVals, _parsedSqlStatement.Model);
+            }
+
             AttribDefs.DbInstance.TrySetValue(attribVals, DatabaseName);
             AttribDefs.PeerAddress.TrySetValue(attribVals, $"{Host}:{PortPathOrId}");
             AttribDefs.PeerHostname.TrySetValue(attribVals, Host);