+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/NetStandardLibraries/CosmosDB/SalesOrder.cs
@@ -1,0 +1,38 @@
+﻿// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+namespace MultiFunctionApplicationHelpers.NetStandardLibraries.CosmosDB
+{
+    using System;
+    using Newtonsoft.Json;
+
+    public class SalesOrder
+    {
+        [JsonProperty(PropertyName = "id")]
+        public string Id { get; set; }
+
+        [JsonProperty(PropertyName = "ponumber")]
+        public string PurchaseOrderNumber { get; set; }
+
+        // used to set expiration policy
+        [JsonProperty(PropertyName = "ttl", NullValueHandling = NullValueHandling.Ignore)]
+        public int? TimeToLive { get; set; }
+
+        public DateTime OrderDate { get; set; }
+        public DateTime ShippedDate { get; set; }
+        public string AccountNumber { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal TaxAmount { get; set; }
+        public decimal Freight { get; set; }
+        public decimal TotalDue { get; set; }
+        public SalesOrderDetail[] Items { get; set; }
+    }
+
+    public class SalesOrderDetail
+    {
+        public int OrderQty { get; set; }
+        public int ProductId { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal LineTotal { get; set; }
+    }
+}